requirements.txt
pytorch==2.0.0
timm==0.6.12
spikingjelly==0.0.0.0.14
scikit-learn==1.3.0
cupy
torchinfo